pyproject.toml
[project]
name = "smoczek"
version = "0.1.0"
description = "tamagochi"
requires-python = ">=3.13"
dependencies = [
    "Pillow"
]
